Introduced the WarehouseItem Class, fixed inheritance, can't figure out how to actually assign the split CSV to a product object, but getting close.
