+++ POS-MidtermTerm-Project1/Product.cs
@@ -8,18 +8,11 @@ namespace POS_MidtermTerm_Project1
 {
     public class Product
     {
+        public int productID { get; set; }
         public string name { get; set; }
-        public string category { get; set; }
         public string description { get; set; }
+        public string category { get; set; }
         public double price { get; set; }
-
-        //internal Product(string productName, string prodcutDescription, 
-        //                 string productCategory, double productPrice)
-        //{
-        //    name = productName;
-        //    category = productCategory;
-        //    description = prodcutDescription;
-        //    price = productPrice;
-        //}
+        public char productChar { get; set; }
     }    
 }
